Add files via upload
Player passing statistics for the big-five leagues aggregated over the 2016/17 and 2017/18 seasons. Players that have played for more than one team in the included leagues during this period have multiple entries e.g. Mohamed Salah is included twice, with separate data for Roma and Liverpool.

More information on the model and analysis is available here: http://2plus2equals11.com/2018/08/12/on-quantifying-passing-skill